refactor(cp): move partial utils/project to core
