+++ src/utils/projects.ts
@@ -1,9 +1,9 @@
 import { select } from "@inquirer/prompts";
-import { commanderInit } from "./commands";
 import { getEnvFilesDirectory } from "./env";
 import { readdir } from "./file";
+import { commanderInit } from "@/core/copy/command";
 
-export type Project = {
+export type Directory = {
 	name: string;
 	value: string;
 };
@@ -11,55 +11,22 @@ export type Project = {
 /**
  * Asynchronously retrieves a list of directories within the envFilesDirectory.
  *
+ * @param directory - The directory within the user's home directory where the environment files directory is located.
  * @returns A promise that resolves to an array of directory names.
  *
  * @throws If there is an issue reading the directory contents.
  *
  * @example
  * // Usage example:
- * const projects = await getProjectsList();
- * console.log(projects); // ['project1', 'project2', ...]
+ * const directories = await getDirectories("/path/to/directory");
+ * console.log(directories);
  */
-export async function getProjectsList(): Promise<Project[]> {
-	const envFilesDirectory = await getEnvFilesDirectory();
-	const dirents = await readdir(envFilesDirectory, { withFileTypes: true });
+export async function getDirectories(directory: string): Promise<Directory[]> {
+	const dirents = await readdir(directory, { withFileTypes: true });
 	return dirents
 		.filter((dirent) => dirent.isDirectory())
 		.map((dirent) => ({
 			name: dirent.name,
 			value: dirent.name,
 		}));
-}
-
-/**
- * Asynchronously prompts the user to select a project from the available list or uses the project specified through command-line options.
- *
- * @param projects - An array of available project names.
- * @returns A promise that resolves to the selected project name.
- *
- * @throws If there is an issue with the user prompt or if the specified project is not found.
- *
- * @example
- * // Usage example:
- * const projects = ["project1", "project2", ...];
- * const selectedProject = await selectProject(projects);
- * console.log(selectedProject); // 'project1' or 'project2'
- */
-export async function selectProject(projects: Project[]): Promise<string> {
-	const options = commanderInit();
-
-	if (!options.project) {
-		const project = await select({
-			message: "Select a project to copy .env files:",
-			choices: projects,
-		});
-		return project;
-	}
-
-	if (!projects.includes(options.project)) {
-		console.log("Error: Specified project not found in the directory.");
-		process.exit(1); // Exit the process with an error code
-	}
-
-	return options.project;
 }
